fix(dist): ship all completions and manpage
- build.rs writes to CARGO_MANIFEST_DIR (source tree) instead of
  OUT_DIR so generated files are packaged by cargo-dist
- Add bash/fish/elvish/powershell completions and manpage to
  Cargo.toml include list
- Deduplicate CLI definition via include!("src/cli_definition.rs")
  shared between build.rs and main.rs; removes hand-written clap
  Command in build.rs that could drift from the derive-based Cli
- Clap derive feature added to build-dependencies
- Compile-time assert guards DEFAULT_LOG_MAX_BYTES sync between
  cli_definition.rs and log.rs
