+++ Cargo.toml
@@ -30,7 +30,7 @@ strsim = "0.11"
 tempfile = "3"
 
 [build-dependencies]
-clap = "4"
+clap = { version = "4", features = ["derive"] }
 clap_complete = "4"
 clap_mangen = "0.2"
 
@@ -51,4 +51,11 @@ cargo-dist-version = "0.28.0"
 ci = ["github"]
 installers = ["shell", "homebrew"]
 targets = ["aarch64-apple-darwin", "x86_64-apple-darwin", "x86_64-unknown-linux-gnu", "aarch64-unknown-linux-gnu"]
-include = ["completions/_zsh-clean-history"]
+include = [
+    "completions/_zsh-clean-history",
+    "completions/zsh-clean-history.bash",
+    "completions/zsh-clean-history.fish",
+    "completions/zsh-clean-history.elv",
+    "completions/_zsh-clean-history.ps1",
+    "man/zsh-clean-history.1",
+]
